// frontend/src/backend.d.ts
export enum EmployeeRole {
    hrManager = "hrManager",
    manager = "manager",
    superAdmin = "superAdmin",
    employee = "employee",
    hrStaff = "hrStaff"
}